// src/common/compoenents/Header/interface.ts
// Generated by https://quicktype.io

export interface Result {
  result: SearchResult;
  code: number;
}

export interface SearchResult {
  albums: Album[];
  artists: Artist[];
  songs: Song[];
  playlists: Playlist[];
  order: string[];
}

export interface Album {
  id: number;
  name: string;
  artist: Artist;
  publishTime: number;
  size: number;
  copyrightId: number;
  status: number;
  picId: number;
  mark: number;
}

export interface Artist {
  id: number;
  name: string;
  picUrl: null | string;
  alias: string[];
  albumSize: number;
  picId: number;
  img1v1Url: string;
  img1v1: number;
  alia?: string[];
  trans: null;
  accountId?: number;
}

export interface Playlist {
  id: number;
  name: string;
  coverImgUrl: string;
  creator: null;
  subscribed: boolean;
  trackCount: number;
  userId: number;
  playCount: number;
  bookCount: number;
  specialType: number;
  officialTags: null;
  description: string;
  highQuality: boolean;
}

export interface Song {
  id: number;
  name: string;
  artists: Artist[];
  album: SongAlbum;
  duration: number;
  copyrightId: number;
  status: number;
  alias: any[];
  rtype: number;
  ftype: number;
  mvid: number;
  fee: number;
  rUrl: null;
  mark: number;
}

export interface SongAlbum {
  id: number;
  name: string;
  artist: Artist;
  publishTime: number;
  size: number;
  copyrightId: number;
  status: number;
  picId: number;
  mark: number;
  alia?: string[];
}